Añadido datos double

+++ src/main/java/ProyectoTP1/CalculoMedia/CalculoMedia.java
@@ -3,12 +3,12 @@ package ProyectoTP1.CalculoMedia;
 public class CalculoMedia {
 
 	public static void main(String[] args) {
-		// Programa que calcule la media de tres números.
+		// Programa sencillo que calcule la media de tres números decimales.
 
-				int media;
-				int a=3;
-				int b=9;
-				int c=13;
+				double media;
+				double a=3;
+				double b=9;
+				double c=13;
 				
 				media=a+b+c/3;
 				System.out.println("La media de "+a+", "+b+" y "+c+" es "+media);
